% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwaa_data.R
\name{nwaa_data}
\alias{nwaa_data}
\title{Download NWAA model output (generic)}
\usage{
nwaa_data(
  model_id,
  variable_ids,
  location_type,
  location_id,
  time_res = "monthly",
  range = c("recent", "historical", "custom"),
  start = NULL,
  end = NULL,
  intersection = NULL,
  skip = 0,
  format = "csv",
  quiet = TRUE
)
}
\arguments{
\item{model_id}{Model ID. See \code{\link{nwaa_models}}.}

\item{variable_ids}{One or more variable IDs. See \code{\link{nwaa_variables}}.}

\item{location_type}{One of \code{\link{nwaa_location_types}}$type.}

\item{location_id}{Identifier for the selected location type.}

\item{time_res}{"monthly", "annualwy", "annualcy".}

\item{range}{"recent", "historical", "custom".}

\item{start, end}{For range = "custom".}

\item{intersection}{Optional, e.g. "overlap" or "envelop".}

\item{skip}{Record offset for paging.}

\item{format}{"csv", "json", or "geojson".}

\item{quiet}{If FALSE, prints request URL and content type.}
}
\description{
Download NWAA model output (generic)
}
