% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\name{nwaa_location}
\alias{nwaa_location}
\title{Build an NWAA location string}
\usage{
nwaa_location(type, id, validate = TRUE)
}
\arguments{
\item{type}{Location type used by NWAA, see nwaa_location_types().}

\item{id}{Location identifier for that type.}

\item{validate}{If TRUE, checks type against nwaa_location_types().}
}
\value{
A string like "type:id".
}
\description{
Build an NWAA location string
}
