% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_codes.R
\name{nwaa_state_codes}
\alias{nwaa_state_codes}
\title{US state codes}
\usage{
nwaa_state_codes()
}
\value{
A tibble with columns \code{state_name} and \code{state_abbr}.
}
\description{
Convenience tibble of US state names and 2-letter abbreviations, sourced
from the base R \code{datasets::state.name} and \code{datasets::state.abb}
vectors.
}
\examples{
nwaa_state_codes()
}
