% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwaa-package.R
\docType{package}
\name{nwaa-package}
\alias{nwaa}
\alias{nwaa-package}
\title{nwaa: 'USGS' 'NWAA' Data Download}
\description{
Download and parse model outputs from the U.S. Geological Survey ('USGS') National Water Availability Assessment ('NWAA') Data Companion web services. Provides typed, validated wrappers for the published NWAA models across the Water Use, Water Quantity, and Integrated Water Availability families, returning results at HUC12 spatial resolution.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://laljeet.github.io/nwaa/}
  \item \url{https://github.com/laljeet/nwaa}
  \item Report bugs at \url{https://github.com/laljeet/nwaa/issues}
}

}
\author{
\strong{Maintainer}: Laljeet Sangha \email{lsanghah2o@gmail.com}

}
\keyword{internal}
