% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwaa-package.R
\docType{package}
\name{nwaa-package}
\alias{nwaa}
\alias{nwaa-package}
\title{nwaa: USGS NWAA Data download}
\description{
Download and parse model outputs from the USGS National Water Availability Assessment (NWAA) Data Companion web services
}
\author{
\strong{Maintainer}: Laljeet Sangha \email{lsanghah2o@gmail.com}

}
\keyword{internal}
