% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{nwaa_get_data}
\alias{nwaa_get_data}
\title{Fetch NWAA data}
\usage{
nwaa_get_data(query, quiet = TRUE)
}
\arguments{
\item{query}{Query list from nwaa_build_query()}

\item{quiet}{If FALSE, prints URL and content-type}
}
\value{
tibble for csv, list for json, sf object for geojson if sf installed
}
\description{
Fetch NWAA data
}
