% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\name{nwaa_location_types}
\alias{nwaa_location_types}
\title{List valid location types used by the NWAA API}
\usage{
nwaa_location_types()
}
\value{
A tibble of location types and notes.
}
\description{
List valid location types used by the NWAA API
}
\examples{
nwaa_location_types()
}
