% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_codes.R
\name{nwaa_state_codes}
\alias{nwaa_state_codes}
\title{Get US state codes (simple built-in mapping)}
\usage{
nwaa_state_codes()
}
\value{
A tibble with state name, abbreviation, and a best-effort FIPS-like code.
}
\description{
Get US state codes (simple built-in mapping)
}
\examples{
nwaa_state_codes()
}
