% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_codes.R
\name{nwaa_statecd}
\alias{nwaa_statecd}
\title{State abbreviations for use with location_type = "statecd"}
\usage{
nwaa_statecd()
}
\value{
A tibble with \code{state_name} and \code{statecd}.
}
\description{
The NWAA API uses lowercase 2-letter state abbreviations (example: "ca", "al").
}
\examples{
nwaa_statecd()
}
