% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wu_catalog.R
\name{nwaa_wu_catalog}
\alias{nwaa_wu_catalog}
\title{Water Use catalog (models, variables, and time ranges)}
\usage{
nwaa_wu_catalog()
}
\value{
A tibble describing Water Use models.
}
\description{
Water Use catalog (models, variables, and time ranges)
}
