% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wu_helpers.R
\name{nwaa_wu_models}
\alias{nwaa_wu_models}
\title{List Water Use models and their available date ranges}
\usage{
nwaa_wu_models()
}
\value{
A tibble with \code{model_id}, \code{model_label}, \code{start_ym}, and \code{end_ym}.
}
\description{
List Water Use models and their available date ranges
}
\examples{
nwaa_wu_models()
}
