% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wu_helpers.R
\name{nwaa_wu_variables}
\alias{nwaa_wu_variables}
\title{List variables for a Water Use model}
\usage{
nwaa_wu_variables(model_id)
}
\arguments{
\item{model_id}{A Water Use model ID from \code{\link{nwaa_wu_models}}.}
}
\value{
A tibble with variable IDs and units.
}
\description{
List variables for a Water Use model
}
\examples{
nwaa_wu_variables("wu-irrigation-wd")
}
